refactor: move to package `jwt`
