+++ pkg/jwt/jwt.go
@@ -1,4 +1,4 @@
-package utils
+package jwt
 
 import (
 	"fmt"
